refactor: split provider progress model
